+++ docs/guides/changelog.md
@@ -12,6 +12,7 @@ description: |-
 - Bump github.com/hashicorp/terraform-plugin-framework from 1.16.0 to 1.16.1 in the terraform-plugin group (PR16)
 - Added Terraform extensions to devcontainer
 - Added `import` capability to all resources and changed resource creation to update existing resource if it already exists (with warning)
+- Added warning when resource deletion fails due to conflict
 
 ## v0.11.4